Attempting update again (some more)
Why tho

+++ Scripts/Services/Seasonal Events/ForsakenFoes/BlackthornDungeonTeleporter.cs
@@ -1,3 +1,5 @@
+using System;
+using Server;
 using Server.Engines.Fellowship;
 using Server.Mobiles;
 
@@ -74,7 +76,7 @@ namespace Server.Items
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write(0); // version
+            writer.Write((int)0); // version
 
             writer.Write(Dest);
             writer.Write((int)Chain);
